+++ CHANGELOG.md
@@ -4,7 +4,7 @@ We follow [Semantic Versions](https://semver.org/).
 
 ## Version next
 
--
+- `enter` now respects overrides
 
 ## Version 0.21.0